+++ lib/rubocop/cop/layout/indent_first_argument.rb
@@ -4,9 +4,8 @@ module RuboCop
   module Cop
     # rubocop:disable Metrics/LineLength
     module Layout
-      # This cop checks the indentation of the first parameter in a method call.
-      # Parameters after the first one are checked by Layout/AlignParameters,
-      # not by this cop.
+      # This cop checks the indentation of the first argument in a method call
+      # or definition.
       #
       # @example
       #
@@ -33,7 +32,7 @@ module RuboCop
       #   second_param
       #
       # @example EnforcedStyle: consistent
-      #   # The first parameter should always be indented one step more than the
+      #   # The first argument should always be indented one step more than the
       #   # preceding line.
       #
       #   # good
@@ -59,8 +58,8 @@ module RuboCop
       #   second_param
       #
       # @example EnforcedStyle: consistent_relative_to_receiver
-      #   # The first parameter should always be indented one level relative to
-      #   # the parent that is receiving the parameter
+      #   # The first argument should always be indented one level relative to
+      #   # the parent that is receiving the argument
       #
       #   # good
       #   some_method(
@@ -85,9 +84,9 @@ module RuboCop
       #   second_params
       #
       # @example EnforcedStyle: special_for_inner_method_call
-      #   # The first parameter should normally be indented one step more than
-      #   # the preceding line, but if it's a parameter for a method call that
-      #   # is itself a parameter in a method call, then the inner parameter
+      #   # The first argument should normally be indented one step more than
+      #   # the preceding line, but if it's a argument for a method call that
+      #   # is itself a argument in a method call, then the inner argument
       #   # should be indented relative to the inner method.
       #
       #   # good
@@ -139,13 +138,13 @@ module RuboCop
       #     nested_first_param),
       #   second_param
       #
-      class FirstParameterIndentation < Cop
+      class IndentFirstArgument < Cop
         # rubocop:enable Metrics/LineLength
         include Alignment
         include ConfigurableEnforcedStyle
         include RangeHelp
 
-        MSG = 'Indent the first parameter one step more than %<base>s.'.freeze
+        MSG = 'Indent the first argument one step more than %<base>s.'.freeze
 
         def on_send(node)
           return if !node.arguments? || node.operator_method?
@@ -163,7 +162,7 @@ module RuboCop
         private
 
         def message(arg_node)
-          return 'Bad indentation of the first parameter.' unless arg_node
+          return 'Bad indentation of the first argument.' unless arg_node
 
           send_node = arg_node.parent
           text = base_range(send_node, arg_node).source.strip