+++ lib/rubocop/cop/metrics/line_length.rb
@@ -24,9 +24,10 @@ module RuboCop
       #   - AlignHash
       #   - AlignParameters
       #   - ClosingParenthesisIndentation
-      #   - FirstParameterIndentation
-      #   - IndentArray
-      #   - IndentHash
+      #   - IndentFirstArgument
+      #   - IndentFirstArrayElement
+      #   - IndentFirstHashElement
+      #   - IndentFirstParameter
       #   - MultilineArrayLineBreaks
       #   - MultilineHashBraceLayout
       #   - MultilineHashKeyLineBreaks