+++ spec/rubocop/cop/layout/empty_lines_around_access_modifier_spec.rb
@@ -297,6 +297,17 @@ RSpec.describe RuboCop::Cop::Layout::EmptyLinesAroundAccessModifier do
       RUBY
     end
 
+    it 'accepts missing blank line when at the end of specifying ' \
+       '`self`' do
+      expect_no_offenses(<<-RUBY.strip_indent)
+        class << self
+          def test; end
+
+          #{access_modifier}
+        end
+      RUBY
+    end
+
     it 'requires blank line when next line started with end' do
       inspect_source(<<-RUBY.strip_indent)
         class Test